+++ NEWS.md
@@ -12,7 +12,7 @@ package must be manually installed to enable parallel processing).
 
 ### Other Improvements
 
-* Improved automatic format detection by `read_chroms`.
+* Improved automatic filetype detection by `read_chroms`.
 * Refactored `read_thermoraw` function to simplify paths.
 * The `read_thermoraw` function now uses temp directory (without prompting) if an export directory is not specified through the `path_out` argument.
 * Added additional tests, attaining 74% test coverage.